==> Section 5. ANN Project Implementation/3. Predictation with trained ANN Model.ipynb <==
import tensorflow as tf
from tensorflow.keras.models import load_model
import pickle
import pandas as pd
import numpy as np

# Load Encoder and Scaler

with open('le_gender.pkl','rb') as f:
    le_gender=pickle.load(f)

with open('ohe_geo.pkl','rb') as f:
    ohe_geo= pickle.load(f)

with open('std_scaler.pkl','rb') as f:
    scaler=pickle.load(f)

input_data = {
    'CreditScore': 600,
    'Geography': 'France',
    'Gender': 'Female',
    'Age': 40,
    'Tenure': 3,
    'Balance': 60000.0,
    'NumOfProducts': 2,
    'HasCrCard': 1,
    'IsActiveMember': 1,
    'EstimatedSalary': 50000.0
}


# Convert input data to data frame
input_df= pd.DataFrame([input_data])
input_df

# Transform Gender Column

input_df['Gender']= le_gender.transform(input_df['Gender'])
input_df

# Transform Geography Column

geo_transformed= ohe_geo.transform(input_df[['Geography']])
geo_transformed

geo_transformed_features_name= ohe_geo.get_feature_names_out(['Geography'])
geo_transformed_features_name

geo_transformed_df= pd.DataFrame(geo_transformed,columns=geo_transformed_features_name)
geo_transformed_df

# Drop original Geography Column and concatenate transformed geography column

input_df= input_df.drop('Geography', axis=1)
input_df=pd.concat([input_df.reset_index(drop=True),geo_transformed_df], axis=1)

input_df

# Apply Standard Scaler

input_scaled= scaler.transform(input_df)
input_scaled

# Load Model

model=load_model('model.h5')

# Make Predictions

prediction= model.predict(input_scaled)
prediction

prediction_probability= prediction[0][0]
prediction_probability

if prediction_probability>0.5:
    print('The customer is likely to churn.')
else:
    print("The customer is not likely to churn.")
