==> Section 5. ANN Project Implementation/4. SalaryRegression.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler, LabelEncoder, OneHotEncoder
import pickle

data=pd.read_csv("Churn_Modelling.csv")
data.head()

### Estimated Salary will be a Output/ Dependant Feature which is continous variable, hence, regression problem

# Drop unnecessary columns
data= data.drop(['RowNumber', 'CustomerId','Surname'],axis=1)
data.head()

# Label Encoder for Gender Column
le_gender= LabelEncoder()
data['Gender']= le_gender.fit_transform(data['Gender'])

# OHE for Geography
ohe_geography= OneHotEncoder(handle_unknown='ignore')
geo_encoded= ohe_geography.fit_transform (data[['Geography']]).toarray()
geo_encoded_df= pd.DataFrame(geo_encoded, columns=ohe_geography.get_feature_names_out(['Geography']))
geo_encoded_df.head()

data= pd.concat([data.drop('Geography', axis=1), geo_encoded_df], axis=1)
data.head()

# Split Data into Dependent and Independant Features

X= data.drop('EstimatedSalary', axis=1)
y= data['EstimatedSalary']

# Train test Split
X_train, X_test,y_train,y_test=train_test_split(X,y,test_size=0.2, random_state=123)

# Scaling Features

std_scaler= StandardScaler()
X_train= std_scaler.fit_transform(X_train)
X_test= std_scaler.transform(X_test)

# Save Encoders and Scaler
with open('le_gender.pkl','wb') as file:
    pickle.dump(le_gender,file)

with open('ohe_geography', 'wb') as file:
    pickle.dump(ohe_geography,file)

with open('std_scaler.pkl','wb') as file:
    pickle.dump(std_scaler, file)

## ANN Regression Problem statement

import tensorflow as tf
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense


# Build Model

model= Sequential([
    Dense(64, activation='relu', input_shape=(X_train.shape[1],)),
    Dense(32, activation='relu'),
    Dense(1) # By Default Linear Activation function will work
])

# Model Compiling

model.compile(optimizer='adam', loss='mean_absolute_error', metrics=['mae'])
model.summary()

## Set up the Tensorboard

from tensorflow.keras.callbacks import EarlyStopping, TensorBoard
import datetime

log_dir= "regressionlogs/fit/" + datetime.datetime.now().strftime("%Y%m%d-%H%M%S")
tf_callback= TensorBoard(log_dir=log_dir, histogram_freq=1)

# Set up early stopping

early_stopping_callback= EarlyStopping(monitor='val_loss', patience=5, restore_best_weights=True)

history= model.fit(
    X_train,y_train,
    validation_data=(X_test, y_test),
    epochs=100,
    callbacks=[early_stopping_callback, tf_callback]
)

## Load tensorBoard Extension

%load_ext tensorboard

%tensorboard --logdir regressionlogs/fit --port 6006

# Evaluate the model on Test Data

test_loss,test_mae= model.evaluate(X_test,y_test)
print(f'Test MAE: {test_mae}')

model.save('regression_model.h5')
